fix: harden final v2 release boundaries (#91)
* fix: harden v2 release boundaries

* fix: verify release packages on clean runners

+++ crates/mint-cli/src/visuals/formatters.rs
@@ -1,6 +1,6 @@
 use std::time::Duration;
 
-pub fn format_bytes(bytes: usize) -> String {
+pub fn format_bytes(bytes: u64) -> String {
     let s = bytes.to_string();
     let mut result = String::new();
     for (i, c) in s.chars().rev().enumerate() {
